- Rename DefaultReviewer to User - PullRequest -> state replaced by  is_declined, is_merged, is_open, is_superseded  - Participant -> role replaced by is_participant, is_reviewer
