+++ atlassian/bitbucket/cloud/repositories/defaultReviewers.py
@@ -3,6 +3,7 @@
 from requests import HTTPError
 
 from ..base import BitbucketCloudBase
+from .users import User
 
 
 class DefaultReviewers(BitbucketCloudBase):
@@ -12,7 +13,7 @@ class DefaultReviewers(BitbucketCloudBase):
     def __get_object(self, data):
         if "errors" in data:
             return
-        return DefaultReviewer(self.url_joiner(self.url, data["display_name"]), data, **self._new_session_args)
+        return User(self.url_joiner(self.url, data["display_name"]), data, **self._new_session_args)
 
     def add(self, user):
         """
@@ -22,7 +23,7 @@ class DefaultReviewers(BitbucketCloudBase):
 
         :param user: string: The user to add
 
-        :return: The added DefaultReviewer object
+        :return: The added User object
         """
         # the mention_id parameter is undocumented but if missed, leads to 400 statuses
         return self.__get_object(self.put(user, data={"mention_id": user}))
@@ -68,33 +69,3 @@ class DefaultReviewers(BitbucketCloudBase):
                 raise
 
         return default_reviewer
-
-
-class DefaultReviewer(BitbucketCloudBase):
-    def __init__(self, url, data, *args, **kwargs):
-        super(DefaultReviewer, self).__init__(url, *args, data=data, expected_type="user", **kwargs)
-
-    @property
-    def display_name(self):
-        return str(self.get_data("display_name"))
-
-    @property
-    def nickname(self):
-        return self.get_data("nickname")
-
-    @property
-    def account_id(self):
-        return self.get_data("account_id")
-
-    @property
-    def uuid(self):
-        return self.get_data("uuid")
-
-    def delete(self):
-        """
-        Deletes the default reviewer
-        """
-        data = super(DefaultReviewer, self).delete(None)
-        if "errors" in data:
-            return
-        return DefaultReviewer(self.url, data, **self._new_session_args)
